feature/refactor-easy-batch
- Refactor EasyBatch
